succesfully fetching to mp api

+++ db/migrate/20201124221323_create_crags.rb
@@ -2,8 +2,8 @@ class CreateCrags < ActiveRecord::Migration[6.0]
   def change
     create_table :crags do |t|
       t.string :name
-      t.integer :lat
-      t.integer :lon
+      t.float :lat
+      t.float :lon
       t.string :location
 
       t.timestamps
